% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun1.R
\name{fun1}
\alias{fun1}
\title{Title}
\usage{
fun1(lm, k)
}
\arguments{
\item{lm}{a linear model}

\item{k}{the number of independent variables in the model}
}
\value{
Shapiro-Wilk Normality Test

Breusch-Pagan Test

Durbin-Watson Test

Variance Inflation Factor

plots of the residuals against each independent variable

a plot of the fitted values vs. residuals

a QQ plot
}
\description{
Title
}
\examples{
y <- rnorm(100, 1, 0)
x1 <- rnorm(100, 2, 1)
x2 <- rnorm(100, 3, 2)
ylm <- lm(y ~ x1 + x2)
fun1(ylm, 2)
}
