% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{Title}
\usage{
myboot(iter, df, alpha = 0.05)
}
\arguments{
\item{iter}{# number of iterations for the bootstrap}

\item{df}{# data grame with first column response, other columns are relevant independent variables}

\item{alpha}{# significance level}
}
\value{
Summary statistics for the bootstrapped betas

Histograms of the bootstrapped betas

Confidence intervals for the boostrapped betas

Classical MLR method point estimates

Classical MLR confidence interval estimates based on alpha
}
\description{
Title
}
\examples{
y <- rnorm(100, 0, 1)
x1 <- rnorm(100, 1, 2)
x2 <- rnorm(100, 2, 3)
data <- cbind(y, x1, x2)
data <- data.frame(data)
myboot(iter=1000, df=data)
}
