% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRODQUAL.R
\docType{data}
\name{PRODQUAL}
\alias{PRODQUAL}
\title{PRODQUAL data set.}
\format{
A data frame with 28 rows and 3 variables:
\describe{
  \item{TEMPERATURE}{degrees Fahrenheit Temperature at which the chemical reaction took place}
  \item{PRESSURE}{Pressure at which the chemical reaction took place in psi}
  \item{QUALITY}{Average of 4 scores from 0-100 independently assigned by four inspectors to measure quality}
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
PRODQUAL
}
\description{
A dataset containing measurements from an experiment on the quality of a product
}
\details{
This is a standard data set from the course MATH 5773 taken from the data sets provided by the text book.
}
\keyword{datasets}
