% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMLR.R
\name{shinyMLR}
\alias{shinyMLR}
\title{3D Regression Plots using shiny}
\usage{
shinyMLR()
}
\value{
3D regression scatterplots with an overlaying plane
}
\description{
User selects a model and a corresponding 3D scatterplot with regression plane is returned
}
\section{shiny}{

The shiny web server will run once this function is invoked and will open a web browser. You should learn how this is implemented.

The web server can be studied \url{https://shiny.rstudio.com/tutorial/}
}

\examples{
\dontrun{ shinyMLR()}
}
