% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testf.R
\name{testf}
\alias{testf}
\title{Test Function}
\usage{
testf(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a vector of squared values in the form of a vector

a string of characters
}
\description{
Takes a vector and returns its square. Also gives a statement as to what the purpose is
}
\examples{
x=1:30; testf(x)
}
