% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.1.R
\name{bayes.1}
\alias{bayes.1}
\title{Bayesian Regression Function}
\usage{
bayes.1(iter = 10000, init = 0.5, y, a, b, df, b.0 = 0, d.0 = 0.001)
}
\arguments{
\item{iter}{number of iterations for the simulations}

\item{init}{prior}

\item{y}{the column number in the data frame for the dependent variable}

\item{a}{the start column number in the data frame for the independent variables}

\item{b}{the end column number in the data frame for the independent variables. The independent variables MUST be consecutive}

\item{df}{data frame for the regression}

\item{b.0}{prior mean of beta. Set to function default, but can be updated by the user}

\item{d.0}{prior scale parameter. Set to function default, but can be updated by the user}
}
\value{
Command-line stats for posterior and mean

Mean for posteriors

Plot of sampler

Summary statistics for classical regression

Confidence intervals for classical regression

Summary statistics for bayesian regression

Posterior plots of marginal distributions

Posterior MCMC Trace and History Plots
}
\description{
Bayesian Regression Function
}
\examples{
\dontrun{data <- diamonds, bayes.1(y = 1, a = 2, b = 3, df = data)}

}
